Break up forest.go
This file was getting huge, so I broke it up into five smaller ones.

Tested: HNC still builds and tests still pass.
